fixed issue where multiple instances of the same Composer class did not get all the correct methods defined on each instance of them. ie:
`first_model.name == "a name"`
`first_composer.name` would return `a name`
if `second_model.name == "a second name"`
`second_composer.name` would still return `a name` if second_composer
is another instance of the same composer as first_composer.
fix is storing the the model methods that were defined on the on the
first instance in a class `attr_accessor` and adding that array to the
list of methods that need to be defined every time on any instance of
that composer class.